U levels.txt: updated with all the waves
To test a wave, simply remove the '//' and this will spawn at the start of the level.
To try another one, simply change the .txt file at any point. After winning or losing, pressing the back arrow will update the levels, and the waves, at which point the changes in the .txt file will take place.