U main: removed 3 second delay when resetting
Now levels are only loaded once at the start. A copy of the current level is made so that the GameData map storing all the levels is not changed. This allows the player to replay any level over and over again while it remains the same (the waves are not destroyed, look level.cpp) and moreover to restart the game without loading the levels from the file thereby saving 3 seconds. The player may reload the levels from the menu in case they wish to change the .txt files.