Level loading (#50)
* U Powerup.h: added clone()

because there was no other way to create a powerup given another powerup

* U Powerup.h: added clone() and moved to .cpp

* U game_data.h: same as game_data.cpp

* U globals.h: added level load game state

* U level: added cloning

* U level.h: same as level.cpp

* U main: removed 3 second delay when resetting

Now levels are only loaded once at the start. A copy of the current level is made so that the GameData map storing all the levels is not changed. This allows the player to replay any level over and over again while it remains the same (the waves are not destroyed, look level.cpp) and moreover to restart the game without loading the levels from the file thereby saving 3 seconds. The player may reload the levels from the menu in case they wish to change the .txt files.

* U game_data: levels are only loaded once

also added a variable (current_level) to hold the current level so the map is never changed

* U Path: idk why this is necessary honestly

* A Powerup.cpp: moved functions to .cpp

 because functions at .h files are gay